+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -4,17 +4,11 @@ import com.sun.jdi.InternalException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.exception.BadRequestException;
-import ru.yandex.practicum.filmorate.exception.ConflictException;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
-import java.time.LocalDate;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,21 +17,12 @@ import java.util.stream.Collectors;
 public class UserService {
 
     private final UserStorage userStorage;
-    private final Map<Integer, User> users = new HashMap<>();
-
 
     public User create(User user) {
-        throwIfUserPrintWrongInfo(user);
-        throwIfUserAlreadyExist(user);
         return userStorage.create(user);
     }
 
     public User update(User user) {
-        throwIfUserPrintWrongInfo(user);
-        if (!users.containsKey(user.getId())) {
-            throw new NotFoundException("HTTP ERROR 404: Невозможно обновить данные, так как пользователя не существует");
-        }
-        throwIfUserAlreadyExist(user);
         return userStorage.update(user);
     }
 
@@ -106,42 +91,5 @@ public class UserService {
                 .filter(friendId -> secondUser.getFriends().contains(friendId))
                 .map(userStorage::findById)
                 .collect(Collectors.toList());
-    }
-
-    void throwIfUserPrintWrongInfo(User user) {
-
-        if (user.getLogin().contains(" ") || user.getLogin().isBlank()) {
-            log.warn("Введенный Логин пользователя: '{}'", user.getLogin());
-            throw new BadRequestException("HTTP ERROR 400: Логин не может быть пустым");
-        }
-
-        if (user.getName() == null || user.getName().equals("")) {
-            user.setName(user.getLogin());
-            log.warn("Не заполнено Имя пользователя заменено на Логин: '{}'", user.getName());
-        }
-
-        if (user.getBirthday().isAfter(LocalDate.now())) {
-            log.warn("Указанная Дата рождения: '{}'", user.getBirthday());
-            throw new BadRequestException("HTTP ERROR 400: Дата рождения не может быть в будущем");
-        }
-
-        if (user.getEmail().isBlank() || user.getEmail() == null || user.getEmail().equals(" ")) {
-            log.warn("Введенный Email пользователя: '{}'", user.getEmail());
-            throw new BadRequestException("HTTP ERROR 400: Email не может быть пустым");
-        }
-    }
-
-    private void throwIfUserAlreadyExist(User userToAdd) {
-        boolean exists = users.values().stream()
-                .anyMatch(user -> isAlreadyExist(userToAdd, user));
-        if (exists) {
-            log.warn("Введенный Email пользователя: '{}'", userToAdd);
-            throw new ConflictException("HTTP ERROR 409: Пользователь с таким Email или логином уже существует");
-        }
-    }
-
-    private boolean isAlreadyExist(User userToAdd, User user) {
-        return userToAdd.getLogin().equals(user.getLogin()) ||
-                userToAdd.getEmail().equals(user.getEmail());
     }
 }
